Move all validation to the client side
