+++ .idea/inspectionProfiles/Project_Default.xml
@@ -192,7 +192,11 @@
       <option name="CHECK_DUPLICATE_KEYS" value="true" />
       <option name="CHECK_DUPLICATE_KEYS_WITH_DIFFERENT_VALUES" value="true" />
     </inspection_tool>
-    <inspection_tool class="DuplicateStringLiteralInspection" enabled="true" level="WARNING" enabled_by_default="true">
+    <inspection_tool class="DuplicateStringLiteralInspection" enabled="true" level="WEAK WARNING" enabled_by_default="true">
+      <scope name="Production" level="WARNING" enabled="true">
+        <option name="MIN_STRING_LENGTH" value="5" />
+        <option name="IGNORE_PROPERTY_KEYS" value="false" />
+      </scope>
       <option name="MIN_STRING_LENGTH" value="3" />
       <option name="IGNORE_PROPERTY_KEYS" value="true" />
     </inspection_tool>
