+++ client/src/main/java/org/spine3/validate/AbstractValidatingBuilder.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.validate;
+package org.spine3.validate;
 
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Message;
@@ -26,8 +26,6 @@ import org.spine3.base.ConversionException;
 import org.spine3.base.FieldPath;
 import org.spine3.base.StringifierRegistry;
 import org.spine3.base.Stringifiers;
-import org.spine3.validate.ConstraintViolation;
-import org.spine3.validate.ConstraintViolationThrowable;
 
 import java.lang.reflect.Type;
 import java.util.List;
