+++ client/src/main/java/org/spine3/validate/MessageFieldValidator.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.validate;
+package org.spine3.validate;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Descriptors.FieldDescriptor;
@@ -28,7 +28,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spine3.base.FieldPath;
 import org.spine3.protobuf.AnyPacker;
-import org.spine3.validate.ConstraintViolation;
 import org.spine3.validate.internal.IfInvalidOption;
 import org.spine3.validate.internal.Time;
 import org.spine3.validate.internal.TimeOption;
