+++ client/src/main/java/org/spine3/validate/MessageValidator.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.validate;
+package org.spine3.validate;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Descriptors.Descriptor;
@@ -26,7 +26,6 @@ import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Message;
 import org.spine3.annotations.Internal;
 import org.spine3.base.FieldPath;
-import org.spine3.validate.ConstraintViolation;
 
 import java.util.List;
