+++ client/src/test/java/org/spine3/validate/IntegerFieldValidatorShould.java
@@ -18,50 +18,43 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.validate;
+package org.spine3.validate;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Any;
-import com.google.protobuf.Descriptors.FieldDescriptor;
+import com.google.protobuf.Int32Value;
+import org.junit.Test;
 import org.spine3.base.FieldPath;
 import org.spine3.protobuf.AnyPacker;
 
-import static java.lang.Math.abs;
-import static org.spine3.protobuf.Values.newDoubleValue;
+import static org.junit.Assert.assertEquals;
 
 /**
- * Validates fields of {@link Double} types.
- *
  * @author Alexander Litus
  */
-class DoubleFieldValidator extends FloatFieldValidatorBase<Double> {
+public class IntegerFieldValidatorShould {
 
-    /**
-     * Creates a new validator instance.
-     *
-     * @param descriptor    a descriptor of the field to validate
-     * @param fieldValues   values to validate
-     * @param rootFieldPath a path to the root field (if present)
-     */
-    DoubleFieldValidator(FieldDescriptor descriptor, ImmutableList<Double> fieldValues, FieldPath rootFieldPath) {
-        super(descriptor, fieldValues, rootFieldPath);
+    private static final Integer VALUE = 2;
+    private static final Integer NEGATIVE_VALUE = -2;
+
+    private final IntegerFieldValidator validator =
+            new IntegerFieldValidator(Any.getDescriptor().getFields().get(0),
+                                      ImmutableList.of(VALUE), FieldPath.getDefaultInstance());
+
+    @Test
+    public void convert_string_to_number() {
+        assertEquals(VALUE, validator.toNumber(VALUE.toString()));
     }
 
-    @Override
-    protected Double toNumber(String value) {
-        final Double min = Double.valueOf(value);
-        return min;
+    @Test
+    public void return_absolute_number_value() {
+        assertEquals(VALUE, validator.getAbs(NEGATIVE_VALUE));
     }
 
-    @Override
-    protected Double getAbs(Double value) {
-        final Double abs = abs(value);
-        return abs;
-    }
-
-    @Override
-    protected Any wrap(Double value) {
-        final Any any = AnyPacker.pack(newDoubleValue(value));
-        return any;
+    @Test
+    public void wrap_to_any() {
+        final Any any = validator.wrap(VALUE);
+        final Int32Value msg = AnyPacker.unpack(any);
+        assertEquals(VALUE, (Integer) msg.getValue());
     }
 }
