+++ client/src/test/proto/spine/test/validate/msg/alt_fields.proto
