+++ server/src/main/java/org/spine3/server/commandbus/Validator.java
@@ -27,8 +27,8 @@ import org.spine3.base.Command;
 import org.spine3.base.CommandContext;
 import org.spine3.base.Commands;
 import org.spine3.server.entity.idfunc.GetTargetIdFromCommand;
-import org.spine3.server.validate.MessageValidator;
 import org.spine3.validate.ConstraintViolation;
+import org.spine3.validate.MessageValidator;
 
 import java.util.List;
