+++ server/src/main/java/org/spine3/server/event/EventBus.java
@@ -36,9 +36,9 @@ import org.spine3.server.event.enrich.EventEnricher;
 import org.spine3.server.outbus.CommandOutputBus;
 import org.spine3.server.outbus.OutputDispatcherRegistry;
 import org.spine3.server.storage.StorageFactory;
-import org.spine3.server.validate.MessageValidator;
 import org.spine3.type.EventClass;
 import org.spine3.validate.ConstraintViolation;
+import org.spine3.validate.MessageValidator;
 
 import javax.annotation.Nullable;
 import java.util.List;
