+++ server/src/test/java/org/spine3/server/event/EventBusShould.java
@@ -34,13 +34,13 @@ import org.spine3.server.command.EventFactory;
 import org.spine3.server.event.enrich.EventEnricher;
 import org.spine3.server.storage.StorageFactory;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
-import org.spine3.server.validate.MessageValidator;
 import org.spine3.test.Tests;
 import org.spine3.test.Tests.MemoizingObserver;
 import org.spine3.test.event.ProjectCreated;
 import org.spine3.test.event.ProjectId;
 import org.spine3.type.EventClass;
 import org.spine3.validate.ConstraintViolation;
+import org.spine3.validate.MessageValidator;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
